+++ src/Symfony/Component/Security/Http/CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
 
  * `UserValueResolver` no longer implements `ArgumentValueResolverInterface`
  * Deprecate calling the constructor of `DefaultLoginRateLimiter` with an empty secret
+ * Add CAS 2.0 access token handler
 
 6.3
 ---